fix file descripter was not passed on OS-X
The error "`recv_io': file descriptor was not passed..." that occurs
on OS-X is mascarading an underling `Broken Pipe` error. The source of
the issue was that the connection is beeing closed prematurely
(sometimes). This is a timing issue and hence the different behavior
when re-running the command or using bundler or not.

The patch blocks until the child sends a confirmation message that
everything is established. This confirmation message must also be sent
when the rails application can't be booted.