Fix readline issues on OS X
The readline issues on OS X seem to be triggered when Ruby's readline
library is built against libedit. I was able to experience the same
issues on Linux by building ruby with libedit rather than GNU readine.

Setting the pgid of the child process, and then requiring readline
seemed to make the process hang. This could be worked around by closing
the stdin of the child process in the Process.spawn call.

However various issues still remained (missing prompt, weird formatting
with pry, etc.)

Eventually I tried just requiring the readline library before changing
the pgid. This seems to solve all issues as far as I can tell.

If this turns out not to work then I think we should detect libedit (via
Readline::VERSION) and swap it out for rb-readline. However this is a
last resort as it may present other compatibility issues / bugs. I'd
rather be able to just use the readline that the user expects so there
are no surprises.

Closes #70

See also #34