Don't watch files outside the root dir
This should be the responsibility of the *user* of the watcher (i.e. the
Application object), not the watcher itself.

If somebody wants to add additional files/directories to the watcher,
which happen to be outside the root dir, then that should be possible -
it's just not the default behaviour of spring.

This also makes the polling watcher much more efficient because it has
to poll many fewer files.

It also reverts the polling watcher implementation to basically what it
was before the listen watcher was added.