Kill the spring server when the Gemfile or Gemfile.lock changes
This ensures that on the next run of a command, the correct bundle is
loaded, or the user sees the correct error message if they have
forgotten to bundle install.

I explored options for making this smarter, such as trying to
automatically restart the server with the new bundle loaded, but it just
got complicated. For example if we did this but there was an error with
the bundle, we'd want to somehow silence that error but show it to the
user on the next run of the command.

In the end I think that changing the bundle happens infrequently enough
that it's acceptable to just kill the server.