Implement a bin/spring binstub
This allows spring to be used without being installed globally (for
example if you use `bundle install --path`). It also provides a
consistent way to invoke spring.

The implementation of bin/spring is such that it prioritises gems
installed in your bundle path, but it will also fall back to gems
install globally. So you can choose whether or not to put spring in your
gem file, and whether or not to use a separate bundle path or install
your bundle globally.

Fixes #220.