Don't remove system-wide gems from the gem path
See #237, bundler might be in the system path.

Instead, I've implemented some basic, er, "parsing" of the lockfile in
order to activate the right version. This is actually an improvement
because now, if the user removes spring from their gemfile, it won't
activate even if their binstubs are still springified.

Closes #237.