Properly fix libedit support
Bug #70 came back because in 8b1a8d6e6c13e75f271af792f666660679e0742d we
started loading the application via Process.spawn rather than via
spawning, and therefore the require of the readline library went back to
coming after the pgid is set.

This time I managed to actually figure out the root cause of the issue.

libedit's implementation of rl_initialize() calls tcsetattr(), which
causes the kernal to generate a SIGTTOU because we cannot write to the
terminal after having set the pgid. The default action of SIGTTOU is to
stop the process, which is why it hangs.

Therefore the proper solution is to just ignore the SIGTTOU.

OMG I am so happy I worked this out.