Abort when bin/spring is called and Spring is already defined
See #252. If you have a Spring 1.0 binstub, but Spring 1.1 on your
system, then the Spring 1.0 binstub will activate Spring 1.1, which will
then try to load your bin/rails, which will then load spring... and on
in on in a loop until you run out of memory. This prevents that and
prompts you to upgrade your binstubs.