fix gems that only add rake tasks
if a gem is in the rake_tasks group, it is supposed to required for
`rake`. Unfortunately, this wasn't working correctly because with
spring the app had already been initialized and had cached all
railties, so the Rakefile require rake_tasks loaded the gems,
but loading of the actual rake tasks couldn't find the railties.

This commit adds a hook for commands to preload things after
config/application has been required (and the main part of
the bundle required), but before the application actually
initializes. Rake command takes advantage of this require the
rake_tasks group