Make the temporary directory path contain uid of the process
This patch resolves #376 a problem with multi-user environment.

On many traditional *nix systems all the users on the same machine share a single global `/tmp`.
If two users on such a system run spring at the same time, the first user's spring process
creates `/tmp/spring` and stores a pidfile and a socket within the directory.
The second users process will then try to store its pidfile within `/tmp/spring`, but it fails
since the directory is owned by another user and (usually) not writable to non-owners.

This patch resolves this problem by making the temporary directory path used by spring
contain the UID of the running process.
If you are the user with UID of 1000, your spring processes will create the directory
`$TMPDIR/spring-1000` and store everything therein.