Configure all threads to not abort on exceptions
This solves long outstanding issue: https://github.com/rails/spring/issues/396.
This happens, because some applications or gems do overwrite global setting of `Thread.abort_on_exception=` which is by default set to `false`.
This leads to application process to exit and making the spring unable to recover from that.