Use Bundler::LockfileParser to parse lockfile instead of regex
The regex that was used to parse the lockfile was not working in the case
where git is configured with `core.autocrlf true`. This changes the line
endings in the lockfile to \r\n, causing the dollar signs in the pattern
to not match.

This changes the parsing of the lockfile to use Bundler::LockfileParser