Run without error if spring in uninstalled group
We have spring in the development and test groups which means that it's
present in the Gemfile but not installed on production. We are also not
using the RAILS_ENV variable when logging in as developer and start a
console with a parameter:

    ./bin/rails console -e production

In this case, a LoadError was raised when trying to load spring. So
since checking the RAILS_ENV is not a reliable detection of spring being
installed, we added a rescue call.

But looking into this issue further, I discovered that we can go through
the list of actually requested gems and skip spring if it wasn't
requested. This solves our specific case. A rescue statement may still
be useful if other people have other reasons for the gem not being
installed but I don't know if that's possible.