Add Spring.spawn_on_env
Currently, the only way Spring can create multiple applications for a
given environment variable value is by changing the `SPRING_APPLICATION_ID`.
This has the tradeoff of not being visible in `bin/spring status` calls,
and being treated as a separate application, when it actually isn't. You
might want to run an application with and without certain features
enabled in the same rails env, and with this patch you can do that and
have Spring treat these as the same app.

Configuration must be set in config/spring_client.rb