Fix reloading in Ruby 3.3
This commit ensures the application is restarted when a preloaded file
is changed. This has been silently broken because the fallback behavior
still results in the application being restarted, but only when the user
next runs a command—hence the acceptance test still passing.

The behavior of `BasicSocket#recv` changed in Ruby 3.3 [1] such that it
now returns `nil` on a closed stream socket, instead of an empty string.
When we call `#empty?` on `nil` the `NoMethodError` is swallowed by the
failsafe thread and the application is not restarted.

This commit fixes the issue by checking both `#nil?` and `#empty?` so it
works with both old and new versions. It also updates the acceptance
test to be more specific about what it considers a "reload" by
asserting the relevant log line exists.

[1]: https://github.com/ruby/ruby/pull/6407