Expect exit status code in spring client
In the previous commit I fixed a scenario where Spring Server failed to pass the application exit code through to Spring Client. Should similar thing happen in future, this can also be detected in Spring Client. It should expect to read some integer and not default to 0 when read nil.

This commit introduces such assertion in Spring Client.

Also fixes #676.

@see 3a8e6096eb129e9b3301dc27986f233e9df5a82f